audit: a second compiler reads our C, and dead code goes (ADR-0248)
Every build has run -Wall -Werror for this project's whole life, so
warning-free looked settled. One gcc -fsyntax-only -Wall -Wextra pass
over the same tree -- green under TCC throughout -- reported ten real
defects:

  - two int/size_t comparisons in containerdef.c, in a file whose own
    other array loops already used size_t
  - four variables declared and never used, one left behind by a
    removal made in the same session
  - a comment block indented as though the if above it guarded what
    followed
  - a struct initialised by position with a field left off the end

None changed behaviour that day, which is what makes them worth
finding: they become bugs on the next edit, and no runtime test reaches
them because there is nothing yet to observe.

test_lint runs gcc in the selftest and a warning fails the build. It is
-fsyntax-only: no object file, nothing that can reach a binary, an
image, a package or the cache, so TCC still compiles every byte this
project ships and ADR-0224 is untouched. The distinction is producer
versus reader.

Proven rather than assumed: with an unused variable injected into
stallwatch.c, TCC compiles it clean under -Wall -Werror and reports
nothing, while test_lint fails and names it.

Also removes the only two genuinely unreferenced functions in the
codebase, sandbox_merge_mutate() and image_recipe_manifest_string() in
pkg.c. An initial scan flagged 310; all but two dissolved once the
generated dispatch table and config.c's cfg_##name token pasting were
accounted for. There are no TODO/FIXME/XXX/HACK markers anywhere.

+++ daemon/src/containerdef.c
@@ -274,7 +274,7 @@ int containerdef_resolve_order(char out_order[][REGISTRY_NAME_MAX])
 void containerdef_write_json_volumes(const struct json_value *root, struct json_writer *w)
 {
 	const struct json_value *jvols;
-	int i;
+	size_t i;
 
 	jw_key(w, "volumes");
 	jw_arr_open(w);
@@ -348,7 +348,7 @@ static void write_stopped_def_json_one(struct container_def *d, struct json_writ
 		const struct json_value *jcmd = json_object_get(root, "cmd");
 
 		if (jcmd != NULL && jcmd->type == JSON_ARRAY) {
-			int k;
+			size_t k;
 
 			for (k = 0; k < jcmd->u.array.count; k++) {
 				const char *s = json_as_string(jcmd->u.array.items[k]);
